+++ packages/provider/src/tests/unit/tasks/powCaptcha/powTasks.unit.test.ts
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 import { stringToHex, u8aToHex } from "@polkadot/util";
-import type { KeyringPair } from "@prosopo/types";
 import {
 	ApiParams,
 	CaptchaStatus,
+	type KeyringPair,
 	POW_SEPARATOR,
 	type PoWChallengeId,
 	type RequestHeaders,
@@ -25,14 +25,13 @@ import type {
 	IProviderDatabase,
 	PoWCaptchaStored,
 } from "@prosopo/types-database";
-import { verifyRecency } from "@prosopo/util";
+import { getIPAddress, verifyRecency } from "@prosopo/util";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { PowCaptchaManager } from "../../../../tasks/powCaptcha/powTasks.js";
 import {
 	checkPowSignature,
 	validateSolution,
 } from "../../../../tasks/powCaptcha/powTasksUtils.js";
-import { getIPAddress } from "../../../../util.js";
 
 vi.mock("@prosopo/util-crypto", () => ({
 	signatureVerify: vi.fn(),