+++ dev/lint/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prosopo/lint
 
+## 2.7.2
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/util@3.0.0
+
 ## 2.7.1
 ### Patch Changes