+++ packages/database/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prosopo/database
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/user-access-policy@3.0.0
+  - @prosopo/types-database@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/config@3.0.0
+
 ## 2.6.9
 ### Patch Changes