+++ packages/dotenv/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @prosopo/dotenv
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/config@3.0.0
+
 ## 2.6.4
 ### Patch Changes